Comments/reports added to server.py
I've put in very basic comments that are outlined by the specification above what I've deem the 'relevant' parts. Not sure what more you think might be necessary but I think anymore would be overkill. 

I've also added reports (a bunch of print statements) after each request as outlined in the specification (so it includes hostname, port, success or failure, the request and if it failed, the reason for doing so. Will need more error checking though.